third_party/blink/web_tests/platform/generic/fast/forms/button/button-type-change-expected.txt

Tests for bug 14439. Button type should be set using el.type.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS btn.type = 'submit'; btn.type is 'submit'
PASS btn.getAttribute('type') is 'submit'
PASS btn.type = 'reset'; btn.type is 'reset'
PASS btn.getAttribute('type') is 'reset'
PASS btn.type = 'button'; btn.type is 'button'
PASS btn.getAttribute('type') is 'button'
PASS btn.type = 'reset'; btn.type is 'reset'
PASS btn.getAttribute('type') is 'reset'
PASS btn.type = 'suBmiT'; btn.type is 'submit'
PASS btn.getAttribute('type') is 'suBmiT'
PASS btn.type = ''; btn.type is 'submit'
PASS btn.getAttribute('type') is ''
PASS btn.type = 'b'; btn.type is 'submit'
PASS btn.getAttribute('type') is 'b'
PASS btn.type = 'RESET'; btn.type is 'reset'
PASS btn.getAttribute('type') is 'RESET'
PASS btn.type = ''; btn.type is 'submit'
PASS btn.getAttribute('type') is ''
PASS btn.type = '/'; btn.type is 'submit'
PASS btn.getAttribute('type') is '/'
PASS btn.type = ' '; btn.type is 'submit'
PASS btn.getAttribute('type') is ' '
PASS btn.type = 'button '; btn.type is 'submit'
PASS btn.getAttribute('type') is 'button '
PASS btn.type = ' b u t t o n '; btn.type is 'submit'
PASS btn.getAttribute('type') is ' b u t t o n '
PASS btn.type = null; btn.type is 'submit'
PASS btn.getAttribute('type') is 'null'
PASS btn.type = undefined; btn.type is 'submit'
PASS btn.getAttribute('type') is 'undefined'
PASS btn.type = 'reset'; txt.value = 'hello'; btn.click(); txt.value is ''
PASS didSubmit is false
PASS btn.type = 'button'; txt.value = 'hello'; btn.click(); txt.value is 'hello'
PASS btn.type = 'submit'; btn.click(); didSubmit is true
PASS successfullyParsed is true

TEST COMPLETE

 
